[#13] ♻️ Refactor: verify 메서드 파라미터 변경
